Disallow user with share token to delete plan

+++ server/TravelService/Controllers/TravelPlansController.cs
@@ -45,6 +45,9 @@ namespace TravelService.Controllers
         [HttpDelete("{id:guid}")]
         public async Task<IActionResult> Delete(Guid id)
         {
+            if (User.IsShareTokenIdentity())
+                return Forbid();
+
             await _service.DeleteAsync(id, User.GetUserId());
             return NoContent();
         }
